可增加商品类别，css，基本功能已实现

+++ ShopProject/LoginModule/src/main/java/com/loginmodule/service/impl/CartServiceImpl.java
@@ -1,12 +1,14 @@
 package com.loginmodule.service.impl;
 
 import com.loginmodule.mapper.CartMapper;
+import com.loginmodule.mapper.UserLogMapper;
 import com.loginmodule.pojo.CartGood;
 import com.loginmodule.pojo.Good;
 import com.loginmodule.service.CartService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
 
@@ -14,11 +16,14 @@ import java.util.Map;
 public class CartServiceImpl implements CartService {
     @Autowired
     CartMapper cartMapper;
+    @Autowired
+    UserLogMapper userLogMapper;
     @Override
     public void insertCart(CartGood cartGood) {
         CartGood cartGood1=cartMapper.selectIfExists(cartGood);
         if(cartGood1==null) {
             cartMapper.insert(cartGood);
+            userLogMapper.insertUserLog(cartGood.getUserId(),cartGood.getGoodId(),2, LocalDateTime.now());
         }
         else{
             cartGood.setNum(cartGood1.getNum()+cartGood.getNum());
@@ -28,7 +33,11 @@ public class CartServiceImpl implements CartService {
 
     @Override
     public void deleteCart(Integer userId,String goodId) {
-        cartMapper.delete(userId,goodId);
+        CartGood cartGood1=cartMapper.selectIfExists(new CartGood(userId,goodId,null));
+        if(cartGood1!=null) {
+            userLogMapper.insertUserLog(userId, goodId, 3,LocalDateTime.now());
+            cartMapper.delete(userId, goodId);
+        }
     }
 
     @Override
